move dockerPush to risu
